requirements.txt
flask==3.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# Dev dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
```